move message.rs to a separate folder
