+++ common/meta/raft-store/src/lib.rs
@@ -14,13 +14,13 @@
 
 pub mod config;
 pub mod log;
+pub mod message;
 pub mod sled_key_spaces;
 pub mod state;
 pub mod state_machine;
+
 // ProtoBuf generated files.
 #[allow(clippy::all)]
 pub mod protobuf {
     tonic::include_proto!("meta");
 }
-
-pub use log::message;
